+++ HISTORY.rst
@@ -7,6 +7,7 @@ Changelog for Libreosteo
 
 - Add zipcode / city autocompletion
 - end of support of python 3.5 and upgrade to Django LTS 2.2
+- Add generic frame for multiple office for the instance. It allows to have multiple office-settings into the product.
 
 
 0.6.0 (2019-10-16)